rtc 自动根据日期计算星期

+++ rtc/driver/rtc.h
@@ -23,7 +23,7 @@ typedef struct {
 }RTCTime_t;
 
 typedef struct {
-	uint16_t year;
+	uint8_t year;
     uint8_t month;
     uint8_t date;
     uint8_t week;
